+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See the [upgrade guide](https://docs.pola.rs/releases/upgrade/0.20/)
 - Changed `count` method to exclude null values
 - Changed `dtype` and `schema` methods to always return instances of data types
 - Added `Enum` type
+- Added `Testing` module
 - Added `sink_ipc`, `sink_csv`, and `sink_ndjson` methods to `LazyFrame`
 - Added `replace` method to `Series` and `Expr`
 - Added `eq`, `eq_missing`, `ne`, and `ne_missing` methods to `Series` and `Expr`